Adding fade out Transition using react-addon-css-transition-group library
Issue: Ran into difficulty when trying to provide an access method for the lists after mutation.

Why: Usage of terniary operators limits us to single line actions.

Fix: Use if-else statements. Pass parent-functions to child and have the child use functions after state mutation.

Due to time constraints and the assignments mentioning of it not needing to function - I'll give the restructuring a good head nod.

The assignment was a fun challenge!